fix issues

+++ cc/util.cc
@@ -1,5 +1,6 @@
 #include "util.h"
 
+#include <cstdarg>
 #include <cstdio>
 #include <iostream>
 
@@ -9,7 +10,7 @@ std::string Format(const char* format, ...) {
 
     while (true) {
         va_start(args, format);
-        int n = vsnprintf((char*)result.data(), result.size(), format, args);
+        size_t n = vsnprintf((char*)result.data(), result.size(), format, args);
         va_end(args);
         if (n <= result.size()) {
             return result;
@@ -22,7 +23,7 @@ std::string Format(const char* format, va_list args) {
     std::string result(256, 0);
 
     while (true) {
-        int n = vsnprintf((char*)result.data(), result.size(), format, args);
+        size_t n = vsnprintf((char*)result.data(), result.size(), format, args);
         if (n <= result.size()) {
             return result;
         }
